Added service methods, REST service controller and unit tests.

+++ src/main/java/hr/lowcostflights/repository/FlightRepository.java
@@ -12,6 +12,6 @@ public interface FlightRepository extends CrudRepository<Flight, Integer> {
 	Flight findOneByOriginAndDestinationAndDepartureDateTimeAndReturnDateTimeAndOutboundStopsAndInboundStopsAndAdultsAndChildrenAndInfantsAndCurrencyAndTotalPrice(
 			Airport origin, Airport destination, LocalDateTime departureDateTime, LocalDateTime returnDateTime,
 			Integer outboundStops, Integer inboundStops, Integer adults, Integer children, Integer infants,
-			String currency, Integer totalPrice);
+			String currency, Double totalPrice);
 	
 }
