+++ src/main/java/hr/lowcostflights/service/FlightService.java
@@ -9,6 +9,6 @@ public interface FlightService {
 
 	public Flight findSpecificFlight(Airport origin, Airport destination, LocalDateTime departureDateTime, LocalDateTime returnDateTime,
 			Integer outboundStops, Integer inboundStops, Integer adults, Integer children, Integer infants,
-			String currency, Integer totalPrice);
+			String currency, Double totalPrice);
 	
 }
